umpire/requirements.txt
docker
docker-compose
aioredis
